Added Homomorphism and Compl/Inter Logic; Removed Analyzer Logic
I added the logic to produce an inverse homomorphic image of a DFA given a homomorphism, and I added the logic to complement a single DFA or to intersect to DFA's. Furthermore, I removed the Analyzer logic because it seems too complicated for the moment. With these changes, all I have to do now is to work out the kinks on my jar compilation for final submission, and I have to add the frontend files for each action to undertake.